Rename folder to be inline with correct conventions.
Fix complete, as of this push, solving the issue where two folders were
created on the repository, with the same name, except for differences in
case.

The repository itself is case sensitive; where as Windows is case
insensitive and will simply pull the folder all into one area. This work
around has fixed the naming issue; changes need to be rectified in the
final files.