Remove redundant .gitignore file.
I've removed the subdirectory .gitignore file, and, have added
a wildcard specification for the subdirectory folders we wanted to
reference inside the arcana folder.

Now, the .gitignore will cascade its effects down into the subdirectory
below it, and will resolve the issues we've been having with version
of Unity issues.