Refactor component codebase.
A big systems update to how the interal structure of the code works.

Of note, there is the addition of the ArcanaObject subclass, that, represents
a large part of how it all works. Every ArcanaObject also holds its
own internal status object that can store a series of states.

Rather than look at states from a simple multiple choice, selection series,
this object actually holds a selection of 'states'. This way, we standardize
the way we treat our components and can safely ensure its compabitable
further down the line.